Extracted TokenBalanceWidgetSettings
